% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_definitions.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize}
\usage{
capitalize(x)
}
\description{
Capitalize initial letter of words in a string.
}
\examples{
capitalize("word1 word2")
}
\keyword{capitalize}
