% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_definitions.R
\docType{data}
\name{final_survey}
\alias{final_survey}
\title{Dataset for races vs tournaments study}
\format{An object of class \code{data.frame} with 125 rows and 12 columns.}
\usage{
data(final_survey)
}
\description{
Final survey data
}
\examples{
data(final_survey)
summary(final_survey)
}
\keyword{contests,}
\keyword{races,}
\keyword{tournaments}
