% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_definitions.R
\name{impute}
\alias{impute}
\title{Impute missing values}
\usage{
impute(x, type = c("random", "zero"))
}
\description{
Impute missing values at random or with a zero.
}
\examples{
impute(c(1, 2, 3, NA), "random")
}
\keyword{imputing,}
\keyword{missing}
\keyword{values}
